Website: Domain and emails are now defined in the config
This should reduce the large number of files users have to edit, and should make it easier to avoid conflicts when pulling updates.
These changes have NOT BEEN TESTED yet. This change should be merged into a new dev branch for testing before being merged into the main branch.

I've also updated the README to mention that the email variables need to be updated.